% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessToken.R
\name{accessToken}
\alias{accessToken}
\alias{setAccessToken}
\title{Get and set GitHub access tokens}
\usage{
accessToken()

setAccessToken(
  token,
  disk.cache = TRUE,
  url = "https://api.github.com",
  app.key = NULL,
  app.secret = NULL,
  user.agent = NULL
)
}
\arguments{
\item{token}{String containing a GitHub personal access token.
This should have the \code{"read:org"} and \code{"read:user"} scopes.
If missing, the user will be prompted to use GitHub's Oauth web application flow to acquire a token.
If \code{NULL}, any existing tokens are cleared from cache.}

\item{disk.cache}{Logical scalar indicating whether to cache the token to disk.
If \code{FALSE}, the token is only cached in memory, which can be helpful for security purposes.}

\item{app.key}{String containing the key for a GitHub Oauth app.
If \code{NULL}, the default \dQuote{ArtifactDB} application is used.}

\item{app.secret}{String containing the secret for a GitHub Oauth app.
If \code{NULL}, the default \dQuote{ArtifactDB} application is used.}

\item{user.agent}{String specifying the user agent for queries to various endpoints.}

\item{github.url}{String containing the URL for the GitHub API.}
}
\value{
\code{setAccessToken} sets the access token and invisibly returns a list containing:
\itemize{
\item \code{token}, a string containing the token.
\item \code{name}, the name of the GitHub user authenticated by the token.
\item \code{expiry}, the Unix time at which the token expires.
}

\code{accessToken} returns the same list, typically retrieved from one of the caches.
If no token was cached, it will call \code{setAccessToken} with default arguments to obtain one.
}
\description{
Get and set GitHub access tokens for authentication to the gypsum API's endpoints.
}
\examples{
if (interactive()) {
    accessToken()
}
}
\author{
Aaron Lun
}
