% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLatest.R
\name{fetchLatest}
\alias{fetchLatest}
\title{Fetch the latest version}
\usage{
fetchLatest(project, asset, url = restUrl(), config = NULL)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{url}{String containing the URL of the gypsum REST API.}

\item{config}{Deprecated and ignored.}
}
\value{
String containing the latest version of the project.
}
\description{
Fetch the latest version of a project's asset.
}
\examples{
fetchLatest("test-R", "basic")

}
\seealso{
\code{\link{refreshLatest}}, to refresh the latest version.
}
\author{
Aaron Lun
}
