% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchMetadataSchema.R
\name{fetchMetadataSchema}
\alias{fetchMetadataSchema}
\title{Fetch one of the known metadata schemas}
\usage{
fetchMetadataSchema(
  name = "bioconductor",
  cache = cacheDirectory(),
  overwrite = FALSE
)
}
\arguments{
\item{name}{String containing the name of the database.}

\item{cache}{String containing the cache directory.
If \code{NULL}, no caching is performed.}

\item{overwrite}{Logical scalar indicating whether to overwrite an existing file in \code{cache}, if one is present.}
}
\value{
String containing a path to the downloaded schema.
}
\description{
Fetch the JSON metadata schema for one of the known database indices.
These indices are generated from the files uploaded to the gypsum backend,
so clients planning to become part of these indices should validate their metadata against the indices' expectations.
}
\examples{
fetchMetadataSchema()
}
\seealso{
\code{\link{validateMetadata}}, to validate metadata against a chosen schema.
}
\author{
Aaron Lun
}
