% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAssets.R
\name{listAssets}
\alias{listAssets}
\title{List assets}
\usage{
listAssets(project, url = restUrl(), config = NULL)
}
\arguments{
\item{project}{String containing the project name.}

\item{url}{String containing the URL of the gypsum REST API.}

\item{config}{Deprecated and ignored.}
}
\value{
Character vector of asset names.
}
\description{
List all assets in a project.
}
\examples{
listAssets("test-R")

}
\author{
Aaron Lun
}
