% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(url = restUrl(), config = NULL)
}
\arguments{
\item{url}{String containing the URL of the gypsum REST API.}

\item{config}{Deprecated and ignored.}
}
\value{
Character vector of project names.
}
\description{
List all projects in the gypsum backent.
}
\examples{
if (interactive()) {
    listProjects()
}

}
\author{
Aaron Lun
}
