% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restUrl.R
\name{restUrl}
\alias{restUrl}
\title{URL for the REST API}
\usage{
restUrl(url)
}
\arguments{
\item{url}{String containing the URL of the REST API.}
}
\value{
If \code{url} is missing, the current setting of the URL is returned.

If \code{url} is provided, it is used replace the current setting of the URL,
and the \emph{previous} setting of the URL is invisibly returned.
}
\description{
Get or set the URL for the gypsum REST API.
}
\examples{
restUrl()
old <- restUrl("https://some-other.rest-api.io") # replace it.
restUrl()
restUrl(old) # setting it back.
}
\author{
Aaron Lun
}
