% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFiles.R
\name{saveFiles}
\alias{saveFiles}
\title{Save a file from a version of a project asset}
\usage{
saveFiles(
  project,
  asset,
  version,
  prefix = NULL,
  destination = NULL,
  cache = cacheDirectory(),
  overwrite = NULL,
  concurrent = 1,
  config = publicS3Config()
)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{prefix}{String containing the prefix for the path.
If provided, files are only downloaded if they have a relative path (i.e., inside the version subdirectory) that starts with this prefix.
If \code{NULL}, all files associated with this version are listed.}

\item{destination}{String containing the path to a destination subdirectory on the local filesystem. 
This is treated as the version's subdirectory regardless of \code{prefix}.
If \code{NULL}, a path inside \code{cache} is used.}

\item{cache}{String containing the path to the cache directory.
Only used if \code{destination=NULL}.}

\item{overwrite}{Logical scalar indicating whether to overwrite an existing file at \code{destination}.
If \code{FALSE} and \code{destination} exists, the download is skipped.
Defaults to \code{!is.null(destination)}, so no extra downloads are performed if the file is already cached.}

\item{concurrent}{Integer specifying the number of concurrent downloads.}

\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}
}
\value{
The file is downloaded to the local file system.
The destination file path is invisibly returned.
}
\description{
Download a file from the gypsum bucket, for a version of an asset of a project.
}
\examples{
out <- saveFiles("test-R", "upload-check", "v1")
list.files(out, recursive=TRUE, all.files=TRUE)

}
\seealso{
\code{\link{saveFile}}, to save a single file.

\code{\link{cacheDirectory}}, for file caching.
}
\author{
Aaron Lun
}
