% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlockProject.R
\name{unlockProject}
\alias{unlockProject}
\title{Unlock a project}
\usage{
unlockProject(project, url = restUrl(), token = accessToken())
}
\arguments{
\item{project}{String containing the project to remove.}

\item{url}{String containing the URL of the gypsum REST API.}

\item{token}{String containing a GitHub access token to authenticate to the gypsum REST API.
The token must refer to a gypsum administrator account.}
}
\value{
\code{NULL} is invisibly returned if the project was successfully removed.
}
\description{
Unlock a project on the gypsum backend.
This is typically caused by a partial upload that was not properly aborted.
}
\examples{
if (interactive()) {
    unlockProject("test-R")
}
}
\seealso{
\code{\link{abortUpload}}, to correctly abort an upload upon failure.
}
\author{
Aaron Lun
}
