% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLatest.R
\name{fetchLatest}
\alias{fetchLatest}
\title{Fetch the latest version}
\usage{
fetchLatest(project, asset, config = publicS3Config())
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}
}
\value{
String containing the latest version of the project.
}
\description{
Fetch the latest version of a project's asset.
}
\examples{
fetchLatest("test-R", "basic")

}
\author{
Aaron Lun
}
