% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchQuota.R
\name{fetchQuota}
\alias{fetchQuota}
\title{Fetch project quota details}
\usage{
fetchQuota(project, url = restUrl(), config = NULL)
}
\arguments{
\item{project}{String containing the project name.}

\item{url}{String containing the URL of the gypsum REST API.}

\item{config}{Deprecated and ignored.}
}
\value{
List containing \code{baseline}, the baseline quota at time zero in bytes;
\code{growth_rate}, the annual growth rate for the quota in bytes;
and \code{year}, the creation year (i.e., time zero) for this project.
}
\description{
Fetch the quota details for a project.
}
\examples{
fetchQuota("test-R")

}
\seealso{
\code{\link{setQuota}}, to set the quota details.
}
\author{
Aaron Lun
}
