% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchUsage.R
\name{fetchUsage}
\alias{fetchUsage}
\title{Fetch project usage details}
\usage{
fetchUsage(project, config = publicS3Config())
}
\arguments{
\item{project}{String containing the project name.}

\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}
}
\value{
Numeric scalar specifying the quota usage for the project, in bytes.
}
\description{
Fetch the quota usage for a project.
}
\examples{
fetchUsage("test-R")

}
\seealso{
\code{\link{refreshUsage}}, to recompute the used quota.
}
\author{
Aaron Lun
}
