% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAssets.R
\name{listAssets}
\alias{listAssets}
\title{List assets}
\usage{
listAssets(project, config = publicS3Config())
}
\arguments{
\item{project}{String containing the project name.}

\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}
}
\value{
Character vector of asset names.
}
\description{
List all assets in a project.
}
\examples{
listAssets("test-R")

}
\author{
Aaron Lun
}
