% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFiles.R
\name{listFiles}
\alias{listFiles}
\title{List files for a version}
\usage{
listFiles(
  project,
  asset,
  version,
  prefix = NULL,
  include.. = TRUE,
  config = publicS3Config()
)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{prefix}{String containing the prefix for the path.
If provided, files are only listed if they have a relative path (i.e., inside the version subdirectory) that starts with this prefix.
If \code{NULL}, all files associated with this version are listed.}

\item{include..}{Logical scalar indicating whether to list files with path components that start with \code{..}.}

\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}
}
\value{
Character vector of relative paths of files associated with the versioned asset.
}
\description{
List the contents of a version of a project asset.
}
\examples{
listFiles("test-R", "upload-check", "v1")

}
\author{
Aaron Lun
}
