% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(config = publicS3Config())
}
\arguments{
\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}
}
\value{
Character vector of project names.
}
\description{
List all projects in the gypsum backent.
}
\examples{
if (interactive()) {
    listProjects()
}

}
\author{
Aaron Lun
}
