% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVersions.R
\name{listVersions}
\alias{listVersions}
\title{List asset versions}
\usage{
listVersions(project, asset, config = publicS3Config())
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}
}
\value{
Character vector of versions.
}
\description{
List all versions of a project asset.
}
\examples{
listVersions("test-R", "upload-check")

}
\author{
Aaron Lun
}
