% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeProject.R
\name{removeProject}
\alias{removeProject}
\title{Remove a project or its children}
\usage{
removeProject(
  project,
  asset = NULL,
  version = NULL,
  url = restUrl(),
  token = accessToken()
)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{asset}{String containing the asset to remove.
If \code{NULL}, all assets of the \code{project} are removed.}

\item{version}{String containing the version of the asset to remove.
If \code{NULL}, all versions of the specified \code{asset} are removed.
Ignored if \code{asset=NULL}.}

\item{url}{String containing the URL of the gypsum REST API.}

\item{token}{String containing a GitHub access token to authenticate to the gypsum REST API.
The token must refer to a gypsum administrator account.}
}
\value{
\code{NULL} is invisibly returned if the project or its contents was successfully removed.
}
\description{
Remove a project, an asset of the project, or a version of a particular asset.
}
\examples{
if (interactive()) {
    createProject("test-R-remove", owners="LTLA")
    removeProject("test-R-remove")
}
}
\seealso{
\code{\link{createProject}}, to create a project.
}
\author{
Aaron Lun
}
