% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFile.R
\name{saveFile}
\alias{saveFile}
\title{Save a file from a version of a project asset}
\usage{
saveFile(
  project,
  asset,
  version,
  path,
  destination = NULL,
  config = publicS3Config(),
  check = TRUE
)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{path}{String containing the relative path to the file inside the version.}

\item{destination}{String containing the destination path for the file.
If \code{NULL}, a path inside the \code{\link{cacheDirectory}} is used.}

\item{config}{Configuration object for the S3 bucket, see \code{\link{publicS3Config}} for details.}

\item{check}{Whether to check if the file exists before attempting a download.
This may be omitted if the file is known to exist.}
}
\value{
The file is downloaded to the local file system.
The destination file path is invisibly returned.
}
\description{
Download a file for a version of a project from the gypsum bucket.
}
\examples{
out <- saveFile("test-R", "upload-check", "v1", "blah.txt")
readLines(out)

}
\seealso{
\code{\link{savePrefix}}, to save all files with the same prefix.

\code{\link{cacheDirectory}}, for file caching.
}
\author{
Aaron Lun
}
